requirements.txt
supabase
python-dotenv
google-genai
